Extract map_pointer callbacks
